Redirect to 404 when user logged in but no access
The reason we don't redirect to wp-login.php in this scenario is because
then we have two options:

1) Enter an infinite redirect loop thanks to ?redirect_to=
2) Or handle that case, but then we force the user to log in again

Instead we redirect to a 404, which is more in line with "private
content" behaviour, although in this case they can actually see the
content exists.

Eventually this'd be redirecting to a custom page instead.